+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - Add ProtonUp-Qt for managing Proton.
 - Remove ge-install-manager. The project has been archived.
 - Change Proton GE version from Proton-6.5-GE-2 to GE-Proton7-8.
+- Add PeaZip for an archive/compression utility.
 
 ## 2.2.0